+++ langchain-core/src/runnables/tests/runnable_stream_events_v2.test.ts
@@ -24,7 +24,7 @@ import {
   HumanMessage,
   SystemMessage,
 } from "../../messages/index.js";
-import { DynamicStructuredTool, DynamicTool, tool } from "../../tools.js";
+import { DynamicStructuredTool, DynamicTool , tool } from "../../tools.js";
 import { Document } from "../../documents/document.js";
 import { PromptTemplate } from "../../prompts/prompt.js";
 import { GenerationChunk } from "../../outputs.js";
